rework popper component to just be a hook
