+++ src/components/Popper/index.ts
@@ -1,2 +1,1 @@
-export * from './Popper';
 export * from './usePopper';
